feat: 通知ページをコンポーネント化
