+++ src/lib/components/Notifications/Notifications.svelte
@@ -1,0 +1,55 @@
+<script lang="ts">
+	import AwaitNote from '$lib/components/Note/AwaitNote.svelte';
+	import Profile from '$lib/components/Profile/Profile.svelte';
+	import { pubkey } from '$lib/data/setting';
+	import { subscribe } from '$lib/utils/nostr';
+	import Icon from '@iconify/svelte';
+	import type { Event } from 'nostr-tools';
+	import { onMount } from 'svelte';
+	import { writable } from 'svelte/store';
+
+	// Map<通知先id, Map<id,[pubkey,content(emoji)]>>
+	const items = writable(new Map<string, Map<string, [string, string]>>());
+
+	onMount(() => {
+		const sub = subscribe({ kinds: [7], '#p': [$pubkey], limit: 20 });
+
+		sub.on('event', (event: Event) => {
+			const noticeForEvent = event.tags.find(([type]) => type === 'e');
+			if (!noticeForEvent) return;
+			const [, id] = noticeForEvent;
+			items.update((item) => {
+				const current = item.get(id) || new Map<string, [string, string]>();
+				return item.set(
+					id,
+					current.set(event.id, [event.pubkey, event.content])
+				);
+			});
+		});
+	});
+</script>
+
+<h2 class="text-2xl">通知</h2>
+<div class="flex flex-col mt-8">
+	{#each [...$items.entries()] as [id, item]}
+		<div class="p-4 mt-4 border">
+			{#if item}
+				<div class="flex gap-4">
+					{#each [...item.values()] as [pubkey, content]}
+						<div class="flex gap-2 items-center">
+							<Profile npubHex={pubkey} imageOnly={true} />
+							<div class="text-3xl">
+								{#if content === '+'}
+									<Icon icon="mdi:cards-heart" color={'red'} />
+								{:else}
+									{content}
+								{/if}
+							</div>
+						</div>
+					{/each}
+				</div>
+			{/if}
+			<AwaitNote {id} />
+		</div>
+	{/each}
+</div>
